requirements.txt
Django==1.11.12
djangorestframework==3.8.2
